+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to Merlin will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+### Fixed
+- Re-enabled distributed integration testing. Added additional examination to distributed testing.
 
 ### Changed
 - CI now splits linting and testing into different tasks for better utilization of